Server/Server/requirements.txt
astroid==2.2.5
asyncio==3.4.3
blis==0.2.4
boto==2.49.0
boto3==1.9.224
botocore==1.12.224
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
colorzero==1.1
cymem==2.0.2
docutils==0.15.2
gensim==3.8.0
gpiozero==1.5.1
idna==2.8
isort==4.3.21
jmespath==0.9.4
lazy-object-proxy==1.4.1
lda==1.1.0
mccabe==0.6.1
murmurhash==1.0.2
numpy==1.17.0
pbr==3.1.1
Pillow==6.1.0
Pillow-PIL==0.1.dev0
pip==19.2.1
plac==0.9.6
preshed==2.0.1
pylint==2.3.1
python-dateutil==2.8.0
requests==2.22.0
s3transfer==0.2.1
scipy==1.3.1
setuptools==40.8.0
six==1.12.0
smart-open==1.8.4
spacy==2.1.8
srsly==0.1.0
thinc==7.0.8
tqdm==4.35.0
typed-ast==1.4.0
urllib3==1.25.3
wasabi==0.2.2
websockets==8.0.2
wrapt==1.11.2
